Oppdatert mappe navn
