+++ apps/oppsummeringsdokument-service/build.gradle
@@ -1,7 +1,7 @@
 plugins {
     id 'java'
     id "org.sonarqube" version "3.5.0.2730"
-    id 'org.springframework.boot' version "2.7.5"
+    id 'org.springframework.boot' version "2.5.7"
     id 'io.spring.dependency-management' version "1.1.0"
 }
