--- apps/oppsummeringsdokument-service/src/main/java/no/nav/registre/testnorge/oppsummeringsdokumentservice/OppsummeringsdokumentServiceApplicationStarter.java
@@ -1,13 +1,0 @@
-package no.nav.registre.testnorge.oppsummeringsdokumentservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class OppsummeringsdokumentServiceApplicationStarter {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OppsummeringsdokumentServiceApplicationStarter.class, args);
-    }
-}
