--- apps/oppsummeringsdokument-service/src/main/java/no/nav/registre/testnorge/oppsummeringsdokumentservice/repository/model/AvvikModel.java
@@ -1,12 +1,0 @@
-package no.nav.registre.testnorge.oppsummeringsdokumentservice.repository.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AvvikModel {
-    private String id;
-    private String navn;
-    private String alvorlighetsgrad;
-}
