--- apps/oppsummeringsdokument-service/src/main/java/no/nav/registre/testnorge/oppsummeringsdokumentservice/repository/model/FartoeyModel.java
@@ -1,12 +1,0 @@
-package no.nav.registre.testnorge.oppsummeringsdokumentservice.repository.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class FartoeyModel {
-    String skipsregister;
-    String skipstype;
-    String fartsomraade;
-}
