--- apps/oppsummeringsdokument-service/src/main/java/no/nav/registre/testnorge/oppsummeringsdokumentservice/util/WebClientFilter.java
@@ -1,14 +1,0 @@
-package no.nav.registre.testnorge.oppsummeringsdokumentservice.util;
-
-import lombok.experimental.UtilityClass;
-import org.springframework.web.reactive.function.client.WebClientResponseException;
-
-@UtilityClass
-public class WebClientFilter {
-
-    public static boolean is5xxException(Throwable throwable) {
-
-        return throwable instanceof WebClientResponseException wce &&
-                wce.getStatusCode().is5xxServerError();
-    }
-}
