+++ apps/oppsummeringsdokument-service/src/main/java/no/nav/testnav/apps/oppsummeringsdokumentservice/OppsummeringsdokumentServiceApplicationStarter.java
@@ -1,0 +1,13 @@
+package no.nav.testnav.apps.oppsummeringsdokumentservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class OppsummeringsdokumentServiceApplicationStarter {
+
+    public static void main(String[] args) {
+        SpringApplication.run(OppsummeringsdokumentServiceApplicationStarter.class, args);
+    }
+}
