+++ apps/oppsummeringsdokument-service/src/main/java/no/nav/testnav/apps/oppsummeringsdokumentservice/config/ApplicationConfig.java
@@ -1,0 +1,19 @@
+package no.nav.testnav.apps.oppsummeringsdokumentservice.config;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+import no.nav.testnav.libs.servletcore.config.ApplicationCoreConfig;
+import no.nav.testnav.libs.servletsecurity.config.SecureOAuth2ServerToServerConfiguration;
+
+@Slf4j
+@Configuration
+@Import(value = {
+        ApplicationCoreConfig.class,
+        SecureOAuth2ServerToServerConfiguration.class
+})
+@RequiredArgsConstructor
+public class ApplicationConfig {
+}
