+++ apps/oppsummeringsdokument-service/src/main/java/no/nav/testnav/apps/oppsummeringsdokumentservice/config/credentials/AaregSyntServiceProperties.java
@@ -1,0 +1,11 @@
+package no.nav.testnav.apps.oppsummeringsdokumentservice.config.credentials;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+import no.nav.testnav.libs.securitycore.domain.ServerProperties;
+
+@Configuration
+@ConfigurationProperties(prefix = "consumers.aareg-synt-services")
+public class AaregSyntServiceProperties extends ServerProperties {
+}
