+++ apps/oppsummeringsdokument-service/src/main/java/no/nav/testnav/apps/oppsummeringsdokumentservice/repository/model/AvvikModel.java
@@ -1,0 +1,12 @@
+package no.nav.testnav.apps.oppsummeringsdokumentservice.repository.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AvvikModel {
+    private String id;
+    private String navn;
+    private String alvorlighetsgrad;
+}
