+++ apps/oppsummeringsdokument-service/src/main/java/no/nav/testnav/apps/oppsummeringsdokumentservice/repository/model/FartoeyModel.java
@@ -1,0 +1,12 @@
+package no.nav.testnav.apps.oppsummeringsdokumentservice.repository.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class FartoeyModel {
+    String skipsregister;
+    String skipstype;
+    String fartsomraade;
+}
