+++ apps/oppsummeringsdokument-service/src/test/resources/application-test.properties
@@ -1,5 +1,5 @@
 spring.cloud.vault.token=dummy
-ELASTIC_USERNAME=username
-ELASTIC_PASSWORD=password
-ELASTIC_HOST=localhost
-ELASTIC_PORT=9200
+OPENSEARCH_USERNAME=username
+OPENSEARCH_PASSWORD=password
+OPENSEARCH_HOST=localhost
+OPENSEARCH_PORT=9200
